Moved non-expression functions from type_utils to where they are used since they werent widely used in the program, added type checking for operation evaluation
